feat: Refactor TMASetupProvider and enhance user data handling in Home component
