+++ src/app/context/tma.tsx
@@ -1,37 +1,9 @@
 "use client";
 
-import { useEffect, useMemo } from "react";
-import { initMiniApp, postEvent, useInitData, useLaunchParams } from "@telegram-apps/sdk-react";
-import fetchy from "@/lib/fetchy";
-import { useStore } from "@/lib/store";
+import { useEffect } from "react";
+import { initMiniApp, postEvent } from "@telegram-apps/sdk-react";
 
 export default function TMASetupProvider({ children }: { children: React.ReactNode }) {
-    const initLaunchParams = useLaunchParams().initData;
-    const launchParams = useLaunchParams();
-    const initData = useInitData();
-    const { setUserID } = useStore();
-
-    const authData = useMemo(() => {
-        return initLaunchParams || initData;
-    }, [initLaunchParams, initData]);
-
-    useEffect(() => {
-        const { username, id } = authData?.user || {};
-
-        async function addUserToDb() {
-            await fetchy.post("/api/user", {
-                id: id,
-                username: username,
-                data: {
-                    authdata: authData,
-                    launchparam: launchParams,
-                },
-            });
-        }
-        addUserToDb();
-        setUserID(id?.toString() || "");
-    }, [authData, launchParams, setUserID]);
-
     useEffect(() => {
         const [miniApp] = initMiniApp();
